window: add getHandle method to obtain an opaque pointer to the impl window
Before, the & operator was overridden to make it return the GLFWWindow* pointer.
In order to hide the underlying window implementation (GLFW) from the user, GLFW headers must
not be included in sogl header files directly.
However, sogl must still provide a way to access the lower level window pointer, which is now done with getHandle.
The user has the responsibility to include GLFW headers and cast the void* to GLFWWindow* in order to use it.

Fixed some Emscripten incompatibility with caps and num lock keys along the way.